backend/sapling-backend/requirements.txt
flask
flask-cors
google-genai
google-auth-oauthlib
google-api-python-client
requests
python-dotenv
